refactor: Rebate 리팩토링(#354)
